SensorExamples/EnergyMeter/py/requirements.txt
RPi.GPIO=="*"
numpy=="*"